requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
blinker==1.7.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
deprecation==2.1.0
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gotrue==2.4.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.24.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.4
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
postgrest==0.16.8
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
traitlets==5.14.3
typing_extensions>=4.12.2
urllib3==2.2.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
